+++ sdk/horizon/streamer.go
@@ -2,68 +2,16 @@ package horizon
 
 import (
 	"context"
-	"fmt"
 	"sync"
 
 	"github.com/stellar/experimental-payment-channels/sdk/agent"
-	"github.com/stellar/experimental-payment-channels/sdk/state"
-	"github.com/stellar/go/amount"
 	"github.com/stellar/go/clients/horizonclient"
 	"github.com/stellar/go/keypair"
 	"github.com/stellar/go/protocols/horizon"
 )
 
-type Horizon struct {
+type Streamer struct {
 	HorizonClient horizonclient.ClientInterface
-}
-
-func (h *Horizon) GetBalance(accountID *keypair.FromAddress, asset state.Asset) (int64, error) {
-	var account horizon.Account
-	account, err := h.HorizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: accountID.Address()})
-	if err != nil {
-		return 0, fmt.Errorf("getting account details of %s: %w", accountID, err)
-	}
-	for _, b := range account.Balances {
-		if b.Asset.Code == asset.Code() || b.Asset.Issuer == asset.Issuer() {
-			balance, err := amount.ParseInt64(account.Balances[0].Balance)
-			if err != nil {
-				return 0, fmt.Errorf("parsing %s balance of %s: %w", asset, accountID, err)
-			}
-			return balance, nil
-		}
-	}
-	return 0, nil
-}
-
-func (h *Horizon) GetSequenceNumber(accountID *keypair.FromAddress) (int64, error) {
-	account, err := h.HorizonClient.AccountDetail(horizonclient.AccountRequest{AccountID: accountID.Address()})
-	if err != nil {
-		return 0, fmt.Errorf("getting account details of %s: %w", accountID, err)
-	}
-	seqNum, err := account.GetSequenceNumber()
-	if err != nil {
-		return 0, fmt.Errorf("getting sequence number of account %s: %w", accountID, err)
-	}
-	return seqNum, nil
-}
-
-func (h *Horizon) SubmitTx(xdr string) error {
-	_, err := h.HorizonClient.SubmitTransactionXDR(xdr)
-	if err != nil {
-		return fmt.Errorf("submitting tx %s: %w", xdr, buildErr(err))
-	}
-	return nil
-}
-
-func buildErr(err error) error {
-	if hErr := horizonclient.GetError(err); hErr != nil {
-		resultString, rErr := hErr.ResultString()
-		if rErr != nil {
-			resultString = "<error getting result string: " + rErr.Error() + ">"
-		}
-		return fmt.Errorf("%w (%v)", err, resultString)
-	}
-	return err
 }
 
 // StreamTx streams transactions that affect the given accounts, sending each
@@ -77,7 +25,7 @@ func buildErr(err error) error {
 // given accounts. At the moment, to reduce complexity and due to limitations in
 // Horizon, it streams all network transactions. See
 // https://github.com/stellar/go/issues/3874.
-func (h *Horizon) StreamTx(cursor string, accounts ...*keypair.FromAddress) (txs <-chan agent.StreamedTransaction, cancel func()) {
+func (h *Streamer) StreamTx(cursor string, accounts ...*keypair.FromAddress) (txs <-chan agent.StreamedTransaction, cancel func()) {
 	// txsCh is the channel that streamed transactions will be written to.
 	txsCh := make(chan agent.StreamedTransaction)
 
@@ -100,7 +48,7 @@ func (h *Horizon) StreamTx(cursor string, accounts ...*keypair.FromAddress) (txs
 	return txsCh, cancel
 }
 
-func (h *Horizon) streamTx(cursor string, txs chan<- agent.StreamedTransaction, cancel <-chan struct{}) {
+func (h *Streamer) streamTx(cursor string, txs chan<- agent.StreamedTransaction, cancel <-chan struct{}) {
 	ctx, ctxCancel := context.WithCancel(context.Background())
 	go func() {
 		<-cancel
