+++ sdk/horizon/streamer_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
-func TestHorizon_StreamTx_longBlock(t *testing.T) {
+func TestStreamer_StreamTx_longBlock(t *testing.T) {
 	client := &horizonclient.MockClient{}
-	h := Horizon{HorizonClient: client}
+	h := Streamer{HorizonClient: client}
 
 	accountA := keypair.MustRandom()
 	client.On(
@@ -70,9 +70,9 @@ func TestHorizon_StreamTx_longBlock(t *testing.T) {
 	assert.False(t, open, "txs channel not closed but should be after cancel called")
 }
 
-func TestHorizon_StreamTx_manyTxs(t *testing.T) {
+func TestStreamer_StreamTx_manyTxs(t *testing.T) {
 	client := &horizonclient.MockClient{}
-	h := Horizon{HorizonClient: client}
+	h := Streamer{HorizonClient: client}
 
 	accountB := keypair.MustRandom()
 	client.On(
